homework2/requirements.txt
Django==4.2.11
djangorestframework==3.16.1
gunicorn==21.2.0